requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
altair==4.1.0
appdirs==1.4.4
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
audioread==2.1.9
backcall==0.2.0
base58==2.1.0
beautifulsoup4==4.9.3
bleach==3.1.5
blinker==1.4
blis==0.7.4
Bottleneck==1.3.2
cachetools==4.2.2
catalogue==2.0.4
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
clang==5.0
click==7.1.2
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.3
configparser==5.0.0
covidcast==0.0.8
crayons==0.4.0
cycler==0.10.0
cymem==2.0.5
darr==0.3.3
decorator==4.4.2
defusedxml==0.6.0
delphi-epidata==0.0.7
Deprecated==1.2.10
discord==1.7.3
discord.py==1.7.3
docopt==0.6.2
efficientnet==1.0.0
entrypoints==0.3
fastai==1.0.64.dev0
fastcore==1.3.20
fastprogress==1.0.0
ffmpeg==1.4
filelock==3.0.12
flatbuffers==1.12
fuzzywuzzy==0.18.0
gast==0.4.0
gcg==0.2.0
gitdb==4.0.5
GitPython==3.1.7
google-auth==1.34.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.39.0
h5py==3.1.0
huggingface-hub==0.0.12
idna==2.10
image-classifiers==1.0.0
imageio==2.9.0
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.17.2
Jinja2==2.11.2
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==2.2.2
jupyterlab-server==1.2.0
jupyterlab-widgets==1.0.0
keras==2.6.0
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
librosa==0.8.1
llvmlite==0.36.0
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
multidict==5.1.0
munch==2.5.0
murmurhash==1.0.5
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.6.2
notebook==6.1.0
notebook-as-pdf==0.3.1
numba==0.53.1
numexpr==2.7.3
numpy==1.19.5
nvidia-ml-py3==7.352.0
oauthlib==3.1.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.0
pandas-alive==2020.5.15
pandocfilters==1.4.2
parso==0.7.1
pathy==0.5.2
pickleshare==0.7.5
Pillow==7.2.0
plotly==5.1.0
pooch==1.4.0
preshed==3.0.5
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.17.3
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.7.4
pydeck==0.6.2
pyee==7.0.2
PyGithub==1.53
Pygments==2.6.1
PyJWT==1.7.1
pykwalify==999.0
pymmwr==0.2.2
pyparsing==2.4.7
PyPDF2==1.26.0
pyppeteer==0.2.2
pyproj==3.1.0
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.19.0
python-Levenshtein==0.12.2
pytz==2020.1
PyWavelets==1.1.1
pywin32==228
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
regex==2021.8.3
requests==2.24.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.2
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
sacremoses==0.0.45
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.11.1
segmentation-models==1.0.1
selenium==3.141.0
semver==2.10.2
Send2Trash==1.5.0
sentencepiece==0.1.96
Shapely==1.7.1
six==1.15.0
smart-open==3.0.0
smmap==3.0.4
sound==0.1.0
sounddevice==0.4.2
SoundFile==0.10.3.post1
soupsieve==2.2.1
spacy==3.0.6
spacy-legacy==3.0.6
srsly==2.4.1
streamlit==0.85.0
styleformer @ git+https://github.com/PrithivirajDamodaran/Styleformer.git@505ca67741d4f4e4e6f1997e3842b30e3f9a21f8
sudo==1.0.0
tenacity==8.0.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
tensorflow-hub==0.12.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
thinc==8.0.6
threadpoolctl==2.1.0
tifffile==2021.7.30
tokenizers==0.10.3
toml==0.10.2
toolz==0.11.1
torch==1.9.0
torchvision==0.10.0
tornado==6.0.4
tqdm==4.48.0
traitlets==4.3.3
transformers==4.9.1
typer==0.3.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.9
validators==0.18.2
wasabi==0.8.2
watchdog==2.1.3
wcwidth==0.2.5
webdriver-manager==3.2.2
webencodings==0.5.1
websockets==8.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wordcloud==1.8.0
wrapt==1.12.1
yarl==1.6.3
zoltpy==0.0.5.dev0
